Fix TTN stdint types and fall-through warn in poll

+++ TheThingsNetwork.h
@@ -113,9 +113,9 @@ public:
   uint8_t getBW();
   uint8_t getCR();
   uint8_t getSF();
-  unsigned long getFrequency(); 	// TODO: fix return type to STDINT
-  unsigned long getRxBW();			// TODO: fix return type to STDINT
-  unsigned long getWatchDogTimer(); // TODO: fix return type to STDINT
+  uint32_t getFrequency();
+  uint32_t getRxBW();
+  uint32_t getWatchDogTimer();
   int8_t getPower();
   int8_t getRSSI();
   int8_t getSNR();
